+++ pyrene/src/components/SimpleTable/SimpleTable.tsx
@@ -7,7 +7,42 @@ import clsx from 'clsx';
 import Loader from '../Loader/Loader';
 import SimpleTableActionList from './SimpleTableActionList';
 import styles from './simpleTable.css';
-import { SimpleTableProps } from './types';
+import { Row, Action } from './types';
+
+export interface SimpleTableProps {
+  /**
+   * Allows the definition of row actions Type: [{ label: [ string ], onClick: [ function ] }, ...]
+   */
+  actions?: Array<Action>,
+  /**
+   * Sets the Table columns.
+   * Type: [{ accessor: ( string | func ) (required), align: string, cellRenderCallback: func, headerName: string, id: string (required), width: number ]
+   */
+  columns: Array<{
+    accessor: string | ((row: Row, rowIndex: number, columnIndex: number) => string | number),
+    align?: string,
+    cellRenderCallback?: (row: Row, rowIndex: number, columnIndex: number) => string | JSX.Element | number,
+    headerName?: string,
+    id: string,
+    width?: number,
+  }>,
+  /**
+   * Sets the Table data displayed in the rows. Type: [ JSON ]
+   */
+  data: Array<Row>,
+  /**
+   * Disables the component and displays a loader inside of it.
+   */
+  loading?: boolean,
+  /**
+   * Called when the user clicks on a row.
+   */
+  onRowClick?: (row: Row) => void,
+  /**
+   * Called when the user double clicks on a row.
+   */
+  onRowDoubleClick?: (row: Row) => void,
+}
 
 /**
  * Simple Tables are used to display tabular data without the overhead of pagination, sorting and filtering.
@@ -65,12 +100,12 @@ const SimpleTable: FunctionComponent<SimpleTableProps> = ({
                   key={column.id.concat(Object.values(valueRow).join('-'))}
                 >
                   <div className={styles.tableCellContent} style={{ textAlign: column.align as any }}>
-                    {(column?.cellRenderCallback?.(valueRow, rowIndex, columnIndex) || valueRow.value) as React.ReactNode}
+                    {column.cellRenderCallback ? column.cellRenderCallback(valueRow, rowIndex, columnIndex) : valueRow.value}
                   </div>
                 </td>
               );
             })}
-            {!loading && data && data.length > 0 && actions && actions.length > 0 && (
+            {!loading && data.length > 0 && actions && actions.length > 0 && (
               <td
                 className={clsx(styles.tableCell, styles.actionCell)}
                 key={`action-${Object.values(row).join('-')}`}
